requirements.txt
pandas
openpyxl

striprtf

webdriver-manager
selenium

pymongo

openai
langchain
torch
git+https://github.com/huggingface/transformers
sentencepiece

git+https://github.com/marella/ctransformers
optimum
auto-gptq

scipy
sentencepiece
accelerate
bitsandbytes
tiktoken
chromadb
huggingface_hub
gpt4all>=2.0.0
python-dotenv

black